+++ src/aplicativo/models/utils/PossuiValor.java
@@ -1,5 +1,5 @@
 package aplicativo.models.utils;
 
 public interface PossuiValor {
-	int getValor();
+	double getValor();
 }
